requirements.txt
Django==1.8.2
Markdown==2.6.2
Pillow==2.9.0
SQLAlchemy==1.0.8
Whoosh==2.7.0
dj-database-url==0.3.0
dj-static==0.0.6
django-haystack==2.4.0
django-markdown-deux==1.0.5
django-postgrespool==0.3.0
django-toolbelt==0.0.1
gunicorn==19.3.0
markdown2==2.3.0
psycopg2==2.6.1
pytz==2015.4
static3==0.6.1
urllib3==1.11
wsgiref==0.1.2
